+++ scico/test/linop/test_linop.py
@@ -14,7 +14,6 @@ import jax
 
 import scico.numpy as snp
 from scico import linop
-from scico.numpy import BlockArray
 from scico.random import randn
 from scico.typing import JaxArray, PRNGKey
 
@@ -105,7 +104,7 @@ def test_binary_op(testobj, operator):
 
     assert isinstance(comp_op, linop.LinearOperator)  # Ensure we don't get a Map
     assert comp_op.input_dtype == testobj.A.dtype
-    np.testing.assert_allclose(comp_mat @ testobj.x, comp_op @ testobj.x, rtol=5e-5)
+    np.testing.assert_allclose(comp_mat @ testobj.x, comp_op @ testobj.x, rtol=0, atol=1e-5)
 
     # linops of different sizes
     with pytest.raises(ValueError):
@@ -507,78 +506,3 @@ def test_operator_norm():
         D = linop.Diagonal(d)
         Dnorm = linop.operator_norm(D)
         assert np.abs(Dnorm - snp.abs(d).max()) < 1e-5
-
-
-class SliceTestObj:
-    def __init__(self, dtype):
-        self.x = snp.zeros((4, 5, 6, 7), dtype=dtype)
-
-
-@pytest.fixture(scope="module", params=[np.float32, np.complex64])
-def slicetestobj(request):
-    yield SliceTestObj(request.param)
-
-
-slice_examples = [
-    np.s_[1:],
-    np.s_[:, 2:],
-    np.s_[..., 3:],
-    np.s_[1:, :-3],
-    np.s_[1:, :, :3],
-    np.s_[1:, ..., 2:],
-]
-
-
-@pytest.mark.parametrize("idx", slice_examples)
-def test_slice_eval(slicetestobj, idx):
-    x = slicetestobj.x
-    A = linop.Slice(idx=idx, input_shape=x.shape, input_dtype=x.dtype)
-    assert (A @ x).shape == x[idx].shape
-
-
-@pytest.mark.parametrize("idx", slice_examples)
-def test_slice_adj(slicetestobj, idx):
-    x = slicetestobj.x
-    A = linop.Slice(idx=idx, input_shape=x.shape, input_dtype=x.dtype)
-    adjoint_test(A)
-
-
-block_slice_examples = [
-    1,
-    np.s_[0:1],
-    np.s_[:1],
-]
-
-
-@pytest.mark.parametrize("idx", block_slice_examples)
-def test_slice_blockarray(idx):
-    x = BlockArray((snp.zeros((3, 4)), snp.ones((3, 4, 5, 6))))
-    A = linop.Slice(idx=idx, input_shape=x.shape, input_dtype=x.dtype)
-    assert (A @ x).shape == x[idx].shape
-
-
-def test_transpose():
-    shape = (1, 2, 3, 4)
-    perm = (1, 0, 3, 2)
-    x, _ = randn(shape)
-    H = linop.Transpose(shape, perm)
-    np.testing.assert_array_equal(H @ x, x.transpose(perm))
-
-    # transpose transpose is transpose inverse
-    np.testing.assert_array_equal(H.T @ H @ x, x)
-
-
-def test_pad():
-    shape = (2, 3, 4)
-    pad = 1
-    x, _ = randn(shape)
-    H = linop.Pad(shape, pad)
-
-    pad_shape = tuple(n + 2 * pad for n in shape)
-    y = snp.zeros(pad_shape)
-    y = y.at[pad:-pad, pad:-pad, pad:-pad].set(x)
-    np.testing.assert_array_equal(H @ x, y)
-
-    # pad transpose is crop
-    y, _ = randn(pad_shape)
-    np.testing.assert_array_equal(H.T @ y, y[pad:-pad, pad:-pad, pad:-pad])