+++ scico/test/optimize/test_pdhg.py
@@ -3,7 +3,7 @@ import numpy as np
 import jax
 
 import scico.numpy as snp
-from scico import functional, linop, loss, random
+from scico import functional, linop, loss, operator, random
 from scico.numpy import BlockArray
 from scico.optimize import PDHG
 
@@ -125,6 +125,28 @@ class TestReal:
         f = loss.SquaredL2Loss(y=self.y, A=A)
         g = (self.λ / 2) * functional.SquaredL2Norm()
         C = linop.MatrixOperator(self.Bmx)
+        pdhg_ = PDHG(
+            f=f,
+            g=g,
+            C=C,
+            tau=τ,
+            sigma=σ,
+            maxiter=maxiter,
+            x0=A.adj(self.y),
+        )
+        x = pdhg_.solve()
+        assert (snp.linalg.norm(self.grdA(x) - self.grdb) / snp.linalg.norm(self.grdb)) < 1e-4
+
+    def test_nlpdhg(self):
+        maxiter = 300
+        τ = 2e-1
+        σ = 2e-1
+        A = linop.Diagonal(snp.diag(self.Amx))
+        f = loss.SquaredL2Loss(y=self.y, A=A)
+        g = (self.λ / 2) * functional.SquaredL2Norm()
+        cfn = lambda x: self.Bmx @ x
+        Cop = operator.operator_from_function(cfn, "Cop")
+        C = Cop(input_shape=self.Bmx.shape[1:])
         pdhg_ = PDHG(
             f=f,
             g=g,